+++ src/GrammarLanguage.ts
@@ -9,8 +9,9 @@ interface AbstractRuntimeProgramConstructorInterface {
 enum GrammarConstantsCompiler {
   stringTemplate = "stringTemplate", // replacement instructions
   indentCharacter = "indentCharacter",
-  listDelimiter = "listDelimiter",
+  catchAllCellDelimiter = "catchAllCellDelimiter",
   openChildren = "openChildren",
+  joinChildrenWith = "joinChildrenWith",
   closeChildren = "closeChildren"
 }
 
@@ -386,9 +387,9 @@ abstract class GrammarBackedNonRootNode extends GrammarBackedNode {
 
   protected _getCompiledLine() {
     const compiler = this.getDefinition()._getCompilerObject()
-    const listDelimiter = compiler[GrammarConstantsCompiler.listDelimiter]
+    const catchAllCellDelimiter = compiler[GrammarConstantsCompiler.catchAllCellDelimiter]
     const str = compiler[GrammarConstantsCompiler.stringTemplate]
-    return str ? TreeUtils.formatStr(str, listDelimiter, this.cells) : this.getLine()
+    return str ? TreeUtils.formatStr(str, catchAllCellDelimiter, this.cells) : this.getLine()
   }
 
   compile() {
@@ -398,15 +399,13 @@ abstract class GrammarBackedNonRootNode extends GrammarBackedNode {
   // todo: remove
   get cells() {
     const cells: jTreeTypes.stringMap = {}
-    this._getGrammarBackedCellArray()
-      .slice(1)
-      .forEach(cell => {
-        if (!cell.isCatchAll()) cells[cell.getCellTypeId()] = cell.getParsed()
-        else {
-          if (!cells[cell.getCellTypeId()]) cells[cell.getCellTypeId()] = []
-          cells[cell.getCellTypeId()].push(cell.getParsed())
-        }
-      })
+    this._getGrammarBackedCellArray().forEach(cell => {
+      if (!cell.isCatchAll()) cells[cell.getCellTypeId()] = cell.getParsed()
+      else {
+        if (!cells[cell.getCellTypeId()]) cells[cell.getCellTypeId()] = []
+        cells[cell.getCellTypeId()].push(cell.getParsed())
+      }
+    })
     return cells
   }
 }
@@ -426,7 +425,7 @@ class GrammarBackedErrorNode extends GrammarBackedNonRootNode {
 
 class GrammarBackedNonTerminalNode extends GrammarBackedNonRootNode {
   // todo: implement
-  protected _getNodeJoinCharacter() {
+  protected _getChildJoinCharacter() {
     return "\n"
   }
 
@@ -434,11 +433,12 @@ class GrammarBackedNonTerminalNode extends GrammarBackedNonRootNode {
     const compiler = this.getDefinition()._getCompilerObject()
     const openChildrenString = compiler[GrammarConstantsCompiler.openChildren] || ""
     const closeChildrenString = compiler[GrammarConstantsCompiler.closeChildren] || ""
+    const childJoinCharacter = compiler[GrammarConstantsCompiler.joinChildrenWith] || "\n"
 
     const compiledLine = this._getCompiledLine()
     const indent = this._getCompiledIndentation()
 
-    const compiledChildren = this.map(child => child.compile()).join(this._getNodeJoinCharacter())
+    const compiledChildren = this.map(child => child.compile()).join(childJoinCharacter)
 
     return `${indent}${compiledLine}${openChildrenString}
 ${compiledChildren}
@@ -1203,7 +1203,7 @@ class GrammarCompilerNode extends TreeNode {
     const types = [
       GrammarConstantsCompiler.stringTemplate,
       GrammarConstantsCompiler.indentCharacter,
-      GrammarConstantsCompiler.listDelimiter,
+      GrammarConstantsCompiler.catchAllCellDelimiter,
       GrammarConstantsCompiler.openChildren,
       GrammarConstantsCompiler.closeChildren
     ]
@@ -1432,7 +1432,7 @@ abstract class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode 
       // Todo: do we want to add new classes to global namespace?
       return vm.runInThisContext(`{
  ${code}
- ${className}
+ global.${className} = ${className}
 }`)
     } catch (err) {
       console.log("Error in code:")
@@ -1444,6 +1444,7 @@ abstract class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode 
   private _importBrowserConstructor(code: string): jTreeTypes.RunTimeNodeConstructor {
     const tempClassName = "tempConstructor" + TreeUtils.getRandomString(30)
     const script = document.createElement("script")
+    // todo: should we namespace things under 1 grammar?
     script.innerHTML = `window.${tempClassName} = ${code}`
     document.head.appendChild(script)
     return (<any>window)[tempClassName]
@@ -1686,18 +1687,27 @@ class NonRootNodeTypeDefinition extends AbstractGrammarDefinitionNode {
   _getExtendsClassName(isCompiled = false): jTreeTypes.javascriptClassPath {
     if (this._amIRoot()) return "jtree.GrammarBackedRootNode"
 
-    const baseTypeNames: any = {}
-    baseTypeNames[GrammarConstants.blobNode] = "BlobNode"
-    baseTypeNames[GrammarConstants.errorNode] = "ErrorNode"
-    baseTypeNames[GrammarConstants.terminalNode] = "TerminalNode"
-    baseTypeNames[GrammarConstants.nonTerminalNode] = "NonTerminalNode"
     const isNonTerminal = this._getFromExtended(GrammarConstants.inScope) || this._getFromExtended(GrammarConstants.catchAllNodeType)
-    const baseNodeTypeName = baseTypeNames[this._getFromExtended(GrammarConstants.baseNodeType)] || (isNonTerminal ? "NonTerminalNode" : "TerminalNode")
-
-    if (!isCompiled) return "jtree." + baseNodeTypeName // todo: this is incorrect but works for legacy stuff.
-
     const extendedDef = <AbstractGrammarDefinitionNode>this._getExtendedParent()
-    return extendedDef ? extendedDef._getGeneratedClassName() : isCompiled ? "jtree.TerminalNode" : "jtree.NonTerminalNode"
+
+    // todo: this is broken for runtime stuff.
+    if (!isCompiled) {
+      if (extendedDef) {
+        // init parent first
+        extendedDef._getConstructorDefinedInGrammar()
+        return `global.${extendedDef._getGeneratedClassName()}`
+      }
+
+      const baseTypeNames: any = {}
+      baseTypeNames[GrammarConstants.blobNode] = "BlobNode"
+      baseTypeNames[GrammarConstants.errorNode] = "ErrorNode"
+      baseTypeNames[GrammarConstants.terminalNode] = "TerminalNode"
+      baseTypeNames[GrammarConstants.nonTerminalNode] = "NonTerminalNode"
+      const baseNodeTypeName = baseTypeNames[this._getFromExtended(GrammarConstants.baseNodeType)] || (isNonTerminal ? "NonTerminalNode" : "TerminalNode")
+      return `jtree.` + baseNodeTypeName // todo: this is incorrect but works for legacy stuff.
+    }
+
+    return extendedDef ? extendedDef._getGeneratedClassName() : isNonTerminal ? "jtree.TerminalNode" : "jtree.NonTerminalNode"
   }
 
   private _getFirstCellHighlightScope() {